+++ packages/ui/src/modal/cancelBid/CancelBidModal.tsx
@@ -18,6 +18,7 @@ import { faGasPump } from '@fortawesome/free-solid-svg-icons'
 import { truncateAddress } from '../../lib/truncate'
 import { ReservoirWallet } from '@reservoir0x/reservoir-sdk'
 import { WalletClient } from 'viem'
+import getChainBlockExplorerUrl from '../../lib/getChainBlockExplorerUrl'
 
 const ModalCopy = {
   title: 'Cancel Offer',
@@ -82,7 +83,6 @@ export function CancelBidModal({
         stepData,
         totalUsd,
         blockExplorerName,
-        blockExplorerBaseUrl,
         cancelOrder,
       }) => {
         const expires = useTimeSince(bid?.expiration)
@@ -212,7 +212,6 @@ export function CancelBidModal({
                           : 'Confirm cancelation in your wallet'
                       }
                       txHashes={stepData?.currentStepItem.txHashes}
-                      blockExplorerBaseUrl={blockExplorerBaseUrl}
                     />
                     {isAttributeOffer && !stepData?.currentStepItem.txHashes && (
                       <Flex justify="center">
@@ -267,23 +266,24 @@ export function CancelBidModal({
                   </Text>
 
                   <Flex direction="column" align="center" css={{ gap: '$2' }}>
-                    {stepData?.currentStepItem?.txHashes?.map(
-                      (txHash, index) => {
-                        const truncatedTxHash = truncateAddress(txHash)
-                        return (
-                          <Anchor
-                            key={index}
-                            href={`${blockExplorerBaseUrl}/tx/${txHash}`}
-                            color="primary"
-                            weight="medium"
-                            target="_blank"
-                            css={{ fontSize: 12 }}
-                          >
-                            View transaction: {truncatedTxHash}
-                          </Anchor>
-                        )
-                      }
-                    )}
+                    {stepData?.currentStepItem?.txHashes?.map((hash, index) => {
+                      const truncatedTxHash = truncateAddress(hash.txHash)
+                      const blockExplorerBaseUrl = getChainBlockExplorerUrl(
+                        hash.chainId
+                      )
+                      return (
+                        <Anchor
+                          key={index}
+                          href={`${blockExplorerBaseUrl}/tx/${hash.txHash}`}
+                          color="primary"
+                          weight="medium"
+                          target="_blank"
+                          css={{ fontSize: 12 }}
+                        >
+                          View transaction: {truncatedTxHash}
+                        </Anchor>
+                      )
+                    })}
                   </Flex>
                 </Flex>
                 <Button