+++ packages/ui/src/modal/collect/mint/MintContent.tsx
@@ -24,7 +24,7 @@ import {
   faWallet,
 } from '@fortawesome/free-solid-svg-icons'
 import QuantitySelector from '../../QuantitySelector'
-import { CollectModalCopy } from '../CollectModal'
+import { CollectCallbackData, CollectModalCopy } from '../CollectModal'
 import { MintImages } from './MintImages'
 import { CollectCheckout } from '../CollectCheckout'
 import SigninStep from '../../SigninStep'
@@ -35,6 +35,8 @@ import { TokenInfo } from '../TokenInfo'
 import { formatNumber } from '../../../lib/numbers'
 import { truncateAddress } from '../../../lib/truncate'
 import { SelectPaymentToken } from '../../SelectPaymentToken'
+import getChainBlockExplorerUrl from '../../../lib/getChainBlockExplorerUrl'
+import { CurrentStepTxHashes } from '../../CurrentStepTxHashes'
 import { ProviderOptionsContext } from '../../../ReservoirKitProvider'
 
 export const MintContent: FC<
@@ -43,6 +45,7 @@ export const MintContent: FC<
     copy: typeof CollectModalCopy
     open: boolean
     setOpen: React.Dispatch<React.SetStateAction<boolean>>
+    onGoToToken?: (data: CollectCallbackData) => any
   }
 > = ({
   collection,
@@ -63,18 +66,20 @@ export const MintContent: FC<
   feeUsd,
   currentChain,
   contract,
+  address,
   hasEnoughCurrency,
   addFundsLink,
   disableJumperLink,
-  blockExplorerBaseUrl,
   transactionError,
   stepData,
   collectStep,
+  contentMode,
   setCollectStep,
   isConnected,
   collectTokens,
   copy,
   setOpen,
+  onGoToToken,
 }) => {
   const providerOptions = useContext(ProviderOptionsContext)
   const pathMap = stepData?.path
@@ -453,6 +458,7 @@ export const MintContent: FC<
                     style={{ height: 32 }}
                   />
                 </Flex>
+                <CurrentStepTxHashes currentStep={stepData?.currentStep} />
                 <Button disabled={true} css={{ mt: '$4', width: '100%' }}>
                   <Loader />
                   {copy.sweepCtaAwaitingApproval}
@@ -511,6 +517,7 @@ export const MintContent: FC<
                         }}
                       />
                     </Box>
+                    <CurrentStepTxHashes currentStep={stepData?.currentStep} />
                     <Button disabled={true} css={{ mt: '$4', width: '100%' }}>
                       <Loader />
                       {copy.sweepCtaAwaitingApproval}
@@ -567,6 +574,7 @@ export const MintContent: FC<
               />
             </Box>
           </Flex>
+          <CurrentStepTxHashes currentStep={stepData?.currentStep} />
           <Button disabled={true} css={{ m: '$4' }}>
             <Loader />
             {copy.mintCtaAwaitingValidation}
@@ -588,7 +596,10 @@ export const MintContent: FC<
             <Text style="h5" css={{ px: '$5' }}>
               Your mint is complete!
             </Text>
-            <MintImages stepData={stepData} contract={contract} />
+            <MintImages
+              stepData={stepData}
+              tokenKind={collection?.contractKind}
+            />
             <Flex align="center" css={{ gap: '$2', px: '$5' }}>
               <Box
                 css={{
@@ -628,12 +639,15 @@ export const MintContent: FC<
                   Array.isArray(item?.txHashes) &&
                   item?.txHashes.length > 0
                 ) {
-                  return item.txHashes.map((txHash, txHashIndex) => {
-                    const truncatedTxHash = truncateAddress(txHash)
+                  return item.txHashes.map((hash, txHashIndex) => {
+                    const truncatedTxHash = truncateAddress(hash.txHash)
+                    const blockExplorerBaseUrl = getChainBlockExplorerUrl(
+                      hash.chainId
+                    )
                     return (
                       <Anchor
                         key={`${itemIndex}-${txHashIndex}`}
-                        href={`${blockExplorerBaseUrl}/tx/${txHash}`}
+                        href={`${blockExplorerBaseUrl}/tx/${hash.txHash}`}
                         color="primary"
                         weight="medium"
                         target="_blank"
@@ -650,9 +664,39 @@ export const MintContent: FC<
             </Flex>
           </Flex>
           <Flex css={{ width: '100%', px: '$4' }}>
-            <Button css={{ width: '100%' }} onClick={() => setOpen(false)}>
-              {copy.mintCtaClose}
-            </Button>
+            {!!onGoToToken ? (
+              <>
+                <Button
+                  onClick={() => {
+                    setOpen(false)
+                  }}
+                  css={{ flex: 1 }}
+                  color="ghost"
+                >
+                  {copy.mintCtaClose}
+                </Button>
+                <Button
+                  style={{ flex: 1 }}
+                  color="primary"
+                  onClick={() => {
+                    onGoToToken({
+                      collectionId: collection?.id,
+                      maker: address,
+                      stepData,
+                      contentMode,
+                    })
+                  }}
+                >
+                  {copy.mintCtaGoToToken.length > 0
+                    ? copy.mintCtaGoToToken
+                    : `View ${successfulMints > 1 ? 'Tokens' : 'Token'}`}
+                </Button>
+              </>
+            ) : (
+              <Button css={{ width: '100%' }} onClick={() => setOpen(false)}>
+                {copy.mintCtaClose}
+              </Button>
+            )}
           </Flex>
         </Flex>
       )}