+++ packages/ui/CHANGELOG.md
@@ -1,4 +1,25 @@
 
+## [v1.19.2-UI](https://github.com/reservoirprotocol/reservoir-kit/commit/9df6c416b4ac46b50881ce29174ae81ff83540eb) (2023-11-21)
+
+* Fix: txHash refactor build error in Cart component [0e8fc880](https://github.com/reservoirprotocol/reservoir-kit/commit/0e8fc8807f79614d4dbca0d65667e7827a7b553b)
+
+## [v1.19.1-UI](https://github.com/reservoirprotocol/reservoir-kit/commit/7172fcf07793496cb934048d5bd22836782e1371) (2023-11-21)
+
+* Fix: improve BuyModal error handling for 1155s and fix client-side exception [3c7565ff](https://github.com/reservoirprotocol/reservoir-kit/commit/3c7565ff1b989de1cac121347cf3d6c02df449bc)
+
+## [v1.19.0-UI](https://github.com/reservoirprotocol/reservoir-kit/commit/cb380a06e5d6b04db133cb19663214b78271e88a) (2023-11-20)
+
+* CollectModal: improve post mint image experience [d7a9a935](https://github.com/reservoirprotocol/reservoir-kit/commit/d7a9a93562b320437fac197679dafe3b4040f545)
+* Fix alwaysIncludeListingCurrency in paymentTokens [78d8fe03](https://github.com/reservoirprotocol/reservoir-kit/commit/78d8fe030b084174edc154e68fa45d187e4099e0)
+* Use seaport v1.5 for oracle orders [280b1b27](https://github.com/reservoirprotocol/reservoir-kit/commit/280b1b27db2c483082b5583d598327e853c66510)
+* CollectModal & BuyModal: Fix expected price sweep bugs [08822d0a](https://github.com/reservoirprotocol/reservoir-kit/commit/08822d0a8d5e63cfe991fe220aedd9c54cc325cc)
+* BidModal, ListModal, EditBidModal, EditListingModal: filter custom on dropdown [3d610e37](https://github.com/reservoirprotocol/reservoir-kit/commit/3d610e37ae1e50027ebd6c8f299904870eabc37c)
+* Respect maxQuantity in case the path quantity is higher than the max [1012217f](https://github.com/reservoirprotocol/reservoir-kit/commit/1012217f5128364e8acee8ec0359b6d9406cfaee)
+* CollectModal: fix maxQuantity calculation [8c322b92](https://github.com/reservoirprotocol/reservoir-kit/commit/8c322b92ad48a04ab4b321bed027695018b2c205)
+
+### Breaking Changes:
+* txHash has now been replaced with an array of txHashes, which contain the transaction hash and the chain id on which it was originated.
+
 ## [v1.18.10-UI](https://github.com/reservoirprotocol/reservoir-kit/commit/339a9c0ccf21e04e75ed2e5d379cfa17b7e4ff4a) (2023-11-16)
 
 * Fix SelectPaymentToken icon aspect ratio [6124eadb](https://github.com/reservoirprotocol/reservoir-kit/commit/6124eadbbaddbe6d0fc279f2a6aa4195fb6dc0d5)